requirements-mac.txt
pyautogui==0.9.54
pillow==11.1.0
python-dotenv==1.0.1 